refactor: rename internal release and scaffold packages
