+++ internal/cmd/buildctl/shared/files.go
@@ -29,7 +29,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/goplus/spx/v2/internal/releasemeta"
+	"github.com/goplus/spx/v2/internal/release"
 )
 
 var fileDownloadHTTPClient = &http.Client{Timeout: 30 * time.Minute}
@@ -51,9 +51,9 @@ func ensureGoPath() (string, error) {
 }
 
 func defaultRuntimeVersion() (string, error) {
-	version := releasemeta.DefaultReleaseMeta().Runtime.Version
+	version := release.DefaultReleaseMeta().Runtime.Version
 	if version == "" {
-		return "", fmt.Errorf("releasemeta: Runtime.Version is empty")
+		return "", fmt.Errorf("release: Runtime.Version is empty")
 	}
 	return version, nil
 }
