+++ internal/scaffold/go.mod.template
